requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
chromadb
streamlit
pandas
ipykernel
python-dotenv
sentence-transformers


- e.